+++ demo/api/src/documents/links/entities/link.entity.ts
@@ -2,7 +2,9 @@ import {
     BlockDataInterface,
     DocumentInterface,
     PageTreeNodeDocumentEntityScopeService,
+    RootBlock,
     RootBlockDataScalar,
+    RootBlockEntity,
     RootBlockType,
     ScopedEntity,
 } from "@comet/cms-api";
@@ -15,6 +17,7 @@ import { v4 } from "uuid";
 @ObjectType({
     implements: () => [DocumentInterface],
 })
+@RootBlockEntity()
 @ScopedEntity(PageTreeNodeDocumentEntityScopeService)
 export class Link extends BaseEntity implements DocumentInterface {
     [OptionalProps]?: "createdAt" | "updatedAt";
@@ -23,6 +26,7 @@ export class Link extends BaseEntity implements DocumentInterface {
     @Field(() => ID)
     id: string = v4();
 
+    @RootBlock(LinkBlock)
     @Property({ type: new RootBlockType(LinkBlock) })
     @Field(() => RootBlockDataScalar(LinkBlock))
     content: BlockDataInterface;