+++ packages/admin/admin/src/common/toolbar/backbutton/ToolbarBackButton.tsx
@@ -2,6 +2,7 @@ import { ArrowLeft } from "@comet/admin-icons";
 import { type ComponentsOverrides, IconButton as MuiIconButton, type Theme } from "@mui/material";
 import { css, useThemeProps } from "@mui/material/styles";
 import { type ReactNode } from "react";
+import { Link } from "react-router-dom";
 
 import { createComponentSlot } from "../../../helpers/createComponentSlot";
 import { type ThemedComponentBaseProps } from "../../../helpers/ThemedComponentBaseProps";
@@ -13,18 +14,16 @@ export type ToolbarBackButtonClassKey = "root" | "iconButton" | "toolbarItem";
 const Root = createComponentSlot("div")<ToolbarBackButtonClassKey>({
     componentName: "ToolbarBackButton",
     slotName: "root",
-})(
-    ({ theme }) => css`
-        flex: 0;
-        display: flex;
+})(css`
+    flex: 0;
+    display: flex;
 
-        .CometAdminToolbarItem-root {
-            padding: 0;
-        }
-    `,
-);
+    .CometAdminToolbarItem-root {
+        padding: 0;
+    }
+`);
 
-const IconButton = createComponentSlot(MuiIconButton)<ToolbarBackButtonClassKey>({
+const IconButton = createComponentSlot(MuiIconButton<typeof Link>)<ToolbarBackButtonClassKey>({
     componentName: "ToolbarBackButton",
     slotName: "iconButton",
 })(
@@ -52,7 +51,7 @@ export interface ToolbarBackButtonProps
     extends ThemedComponentBaseProps<{
         root: "div";
         toolbarItem: typeof ToolbarItem;
-        iconButton: typeof MuiIconButton;
+        iconButton: typeof MuiIconButton<typeof Link>;
     }> {
     backIcon?: ReactNode;
 }
@@ -69,16 +68,12 @@ export const ToolbarBackButton = (inProps: ToolbarBackButtonProps) => {
         return null;
     }
 
+    const backUrl = stackApi.breadCrumbs[stackApi.breadCrumbs.length - 2].url;
+
     return (
         <Root {...slotProps?.root} {...restProps}>
             <ToolbarItem {...slotProps?.toolbarItem}>
-                <IconButton
-                    onClick={() => {
-                        stackApi?.goBack();
-                    }}
-                    size="large"
-                    {...slotProps?.iconButton}
-                >
+                <IconButton component={Link} to={backUrl} size="large" {...slotProps?.iconButton}>
                     {backIcon}
                 </IconButton>
             </ToolbarItem>