+++ packages/admin/brevo-admin/src/brevoConfiguration/BrevoConfigPage.tsx
@@ -1,0 +1,20 @@
+import { useContentScope } from "@comet/cms-admin";
+import type { JSX } from "react";
+
+import { useBrevoConfig } from "../common/BrevoConfigProvider";
+import { BrevoConfigForm } from "./BrevoConfigForm";
+
+export function BrevoConfigPage(): JSX.Element {
+    const { scopeParts } = useBrevoConfig();
+    const { scope: completeScope } = useContentScope();
+
+    const scope = scopeParts.reduce(
+        (acc, scopePart) => {
+            acc[scopePart] = completeScope[scopePart];
+            return acc;
+        },
+        {} as { [key: string]: unknown },
+    );
+
+    return <BrevoConfigForm scope={scope} />;
+}