+++ packages/admin/brevo-admin/src/emailCampaigns/statistics/EmailCampaignStatistics.gql.ts
@@ -1,0 +1,15 @@
+import { gql } from "@apollo/client";
+
+export const emailCampaignStatistics = gql`
+    query EmailCampaignStatistics($id: ID!) {
+        brevoEmailCampaignStatistics(id: $id) {
+            uniqueClicks
+            unsubscriptions
+            delivered
+            sent
+            softBounces
+            hardBounces
+            uniqueViews
+        }
+    }
+`;