+++ packages/admin/cms-admin/src/dam/DataGrid/DamContextMenu.tsx
@@ -1,8 +1,7 @@
 import { gql, useApolloClient } from "@apollo/client";
-import { RowActionsItem, RowActionsMenu, useEditDialogApi, useErrorDialog, useStackSwitchApi } from "@comet/admin";
+import { downloadFile, RowActionsItem, RowActionsMenu, useEditDialog, useErrorDialog, useStackSwitchApi } from "@comet/admin";
 import { Archive, Delete, Download, Edit, Move, Restore } from "@comet/admin-icons";
 import { Divider } from "@mui/material";
-import { saveAs } from "file-saver";
 import { useState } from "react";
 import { FormattedMessage } from "react-intl";
 
@@ -29,7 +28,7 @@ interface FolderInnerMenuProps {
 }
 
 const FolderInnerMenu = ({ folder, openMoveDialog }: FolderInnerMenuProps) => {
-    const editDialogApi = useEditDialogApi();
+    const [, , editDialogApi] = useEditDialog();
     const errorDialog = useErrorDialog();
     const apolloClient = useApolloClient();
     const { apiUrl } = useCometConfig();
@@ -150,7 +149,7 @@ const FileInnerMenu = ({ file, openMoveDialog }: FileInnerMenuProps) => {
                     <RowActionsItem
                         icon={<Download />}
                         onClick={() => {
-                            saveAs(file.fileUrl, file.name);
+                            downloadFile(file.fileUrl, file.name);
                         }}
                     >
                         <FormattedMessage id="comet.pages.dam.downloadFile" defaultMessage="Download file" />