+++ packages/admin/cms-admin/src/dam/FileForm/previews/ImagePreview.tsx
@@ -1,3 +1,5 @@
+import type { JSX } from "react";
+
 import { ImageCrop } from "../../../common/image/ImageCrop";
 import { type DamFileDetails } from "../EditFile";