+++ packages/api/api-generator/src/commands/generate/utils/__tests__/find-hooks-service/test-entity3-error.ts
@@ -1,0 +1,9 @@
+import { type MutationError } from "@comet/cms-api";
+
+export enum TestEntity3MutationErrorCode {
+    titleTooShort = "titleTooShort",
+}
+export class TestEntity3MutationError implements MutationError {
+    field?: string;
+    code: TestEntity3MutationErrorCode;
+}