+++ packages/eslint-config/nextjs.js
@@ -1,4 +1,4 @@
-import coreConfig from "./core.js";
+import coreConfig, { restrictedImportPatterns } from "./core.js";
 import react from "eslint-plugin-react";
 import globals from "globals";
 import nextPlugin from "@next/eslint-plugin-next";
@@ -11,7 +11,7 @@ export const restrictedImportPaths = [
     {
         name: "next/image",
         importNames: ["default"],
-        message: "Please use Image from @comet/site-nextjs instead",
+        message: "Don't use next/image. See https://docs.comet-dxp.com/docs/faqs/next-image-import-restriction",
     },
 ];
 
@@ -29,6 +29,7 @@ const config = [
                 "error",
                 {
                     paths: restrictedImportPaths,
+                    patterns: restrictedImportPatterns,
                 },
             ],
         },