+++ storybook/src/admin/edit-dialog/EditDialogInRouterTabs.stories.tsx
@@ -33,7 +33,7 @@ const products = [
 ];
 
 type DialogProps = {
-    dialogApiRef: RefObject<IEditDialogApi>;
+    dialogApiRef: RefObject<IEditDialogApi | null>;
 };
 
 const AddProductDialog = ({ dialogApiRef }: DialogProps) => {