+++ packages/admin/admin-icons/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @comet/admin-icons
 
+## 9.0.0-beta.0
+
+### Major Changes
+
+- 5f1566a: Make packages ESM-only
+
+### Minor Changes
+
+- f066335: Add support for React 19
+
+## 8.17.1
+
+### Patch Changes
+
+- 91e9a8f: Update fast-xml-parser to 5.3.6 to fix CVE-2026-26278 (entity expansion DoS)
+
+## 8.17.0
+
+## 8.16.0
+
+## 8.15.0
+
+## 8.14.0
+
+## 8.13.0
+
+## 8.12.0
+
+## 8.11.1
+
+## 8.11.0
+
+### Patch Changes
+
+- 7e34c0b: Fix `Mail` and `Letter` icons
+
 ## 8.10.0
 
 ## 8.9.0